refactor: Clean up the code a little bit
